update theme and add menu toolbar
